Update and delete Method completed for file upload

+++ app/Http/Controllers/ProductController.php
@@ -5,6 +5,7 @@ use App\Http\Requests\StoreProductRequest;
 use Illuminate\Http\Request;
 use App\Models\Product;
 use Illuminate\Support\Facades\Storage;
+use App\Http\Requests\UpdateProductRequest;
 
 
 class ProductController extends Controller
@@ -35,12 +36,12 @@ class ProductController extends Controller
         return response()->json($product, 200);
     }
 
-    public function update(StoreProductRequest $request, Product $product)
+    public function update(UpdateProductRequest $request, Product $product)
     {
         $data = $request->validated();
         if ($request->hasFile('image')) {
             // delete old image
-            if ($product->image) {
+            if ($product->image && Storage::disk('public')->exists($product->image)) {
                 Storage::disk('public')->delete($product->image);
             }
 
@@ -57,7 +58,7 @@ class ProductController extends Controller
 
     public function destroy(Product $product)
     {
-        if ($product->image) {
+        if ($product->image && Storage::disk('public')->exists($product->image)) {
             Storage::disk('public')->delete($product->image);
         }
